GitHub Cleanup
Fork leanKernel and import relevant changes to current working
repository.

+++ buildlean.sh
@@ -1,56 +1,76 @@
 #!/bin/sh
 
-CPU_JOB_NUM=2
-TOOLCHAIN=/root/CodeSourcery/Sourcery_G++_Lite/bin/
-TOOLCHAIN_PREFIX=arm-none-linux-gnueabi-
+CPU_JOB_NUM=8
+TOOLCHAIN_PREFIX=arm-none-eabi-
+
+export USE_CCACHE=1
+export CCACHE_DIR=/Users/TwistedZero/.ccache
+../prebuilt/darwin-x86/ccache/ccache -M 20G
+
+make clean -j$CPU_JOB_NUM
 
 if [ $3 ]; then
-  suffix="test$3"
+suffix="test$3"
 else
-  suffix=""
+suffix=""
 fi
 
 sed -i s/CONFIG_LOCALVERSION=\"-imoseyon-.*\"/CONFIG_LOCALVERSION=\"-imoseyon-${2}AOSP\"/ .config
 
 if [ $1 -eq 2 ]; then
-  sed -i "s/^.*UNLOCK_184.*$/CONFIG_UNLOCK_184MHZ=n/" .config
-  zipfile="imoseyon_leanKernel_v${2}AOSP${suffix}.zip"
+sed -i "s/^.*UNLOCK_184.*$/CONFIG_UNLOCK_184MHZ=n/" .config
+zipfile="imoseyon_leanKernel_v${2}AOSP${suffix}.zip"
 else
-  sed -i "s/^.*UNLOCK_184.*$/CONFIG_UNLOCK_184MHZ=y/" .config
-  zipfile="imoseyon_leanKernel_184Mhz_v${2}AOSP${suffix}.zip"
+sed -i "s/^.*UNLOCK_184.*$/CONFIG_UNLOCK_184MHZ=y/" .config
+zipfile="imoseyon_leanKernel_184Mhz_v${2}AOSP${suffix}.zip"
 fi
 
-make -j$CPU_JOB_NUM ARCH=arm CROSS_COMPILE=$TOOLCHAIN/$TOOLCHAIN_PREFIX
+make -j$CPU_JOB_NUM ARCH=arm CROSS_COMPILE=$TOOLCHAIN_PREFIX
 
 # make nsio module here for now
 cd nsio*
 make
 cd ..
 
-find . -name "*.ko" | xargs $TOOLCHAIN/${TOOLCHAIN_PREFIX}strip --strip-unneeded
+find . -name "*.ko" | xargs ${TOOLCHAIN_PREFIX}strip --strip-unneeded
 
-rm zip.aosp/system/lib/modules/*.ko
+if [ ! -e zip.aosp ]; then
+mkdir zip.aosp
+fi
+if [ ! -e zip.aosp/system ]; then
+mkdir zip.aosp/system
+fi
+if [ ! -e zip.aosp/system/lib ]; then
+mkdir zip.aosp/system/lib
+fi
+if [ ! -e zip.aosp/system/lib/modules ]; then
+mkdir zip.aosp/system/lib/modules
+else
+rm -r zip.aosp/system/lib/modules
+mkdir zip.aosp/system/lib/modules
+fi
 cp drivers/net/wireless/bcm4329/bcm4329.ko zip.aosp/system/lib/modules
 cp drivers/net/tun.ko zip.aosp/system/lib/modules
 cp drivers/staging/zram/zram.ko zip.aosp/system/lib/modules
 cp lib/lzo/lzo_decompress.ko zip.aosp/system/lib/modules
 cp lib/lzo/lzo_compress.ko zip.aosp/system/lib/modules
+if [ ! -e nsio*/*.ko ]; then
 cp nsio*/*.ko zip.aosp/system/lib/modules
+fi
 cp fs/cifs/cifs.ko zip.aosp/system/lib/modules
 cp arch/arm/boot/zImage mkboot.aosp/
 cp .config arch/arm/configs/lean_aosp_defconfig
-cp .config .config.aosp
 
 cd mkboot.aosp
 echo "making boot image"
 ./img.sh
 
 if [ ! $4 ]; then
-	echo "making zip file"
-	cp boot.img ../zip.aosp
-	cd ../zip.aosp
-	rm *.zip
-	zip -r $zipfile *
-	rm /tmp/*.zip
-	cp *.zip /tmp
+echo "making zip file"
+cp boot.img ../zip.aosp
+cd ../zip.aosp
+rm *.zip
+zip -r $zipfile *
+rm /tmp/*.zip
+cp *.zip /tmp
 fi
